[ARF-87799] ARP 21a: Repair 21a e2e tests.
