+++ src/applications/accredited-representative-portal/tests/accreditation/21a/e2e/fixtures/mocks/in-progress-form.json
@@ -1,0 +1,10 @@
+{
+  "data": {
+    "id": "1234",
+    "attributes": {
+      "metadata": {
+        "expiresAt": 9999999999
+      }
+    }
+  }
+}
