+++ src/applications/accredited-representative-portal/tests/accreditation/21a/e2e/fixtures/mocks/user.json
